move stalatest specific classes into another lib
