+++ ledger/ledger-security-evidence/tool/BUILD.bazel
@@ -24,8 +24,8 @@ da_scala_library(
     deps = [
         "//ledger/ledger-api-tests/infrastructure",
         "//ledger/ledger-api-tests/suites",
+        "//libs-scala/security-evidence/scalatest:security-evidence-scalatest",
         "//libs-scala/security-evidence/tag:security-evidence-tag",
-        "//ledger/ledger-security-evidence/lib",
         "@maven//:org_scalatest_scalatest_compatible",
     ],
 )
@@ -40,8 +40,8 @@ da_scala_binary(
     ],
     deps = [
         ":tool",
-        "//libs-scala/security-evidence/tag:security-evidence-tag",
-        "//ledger/ledger-security-evidence/lib",
         "//ledger/sandbox-on-x:sandbox-on-x-it-tests-copy",
+        "//libs-scala/security-evidence/scalatest:security-evidence-scalatest",
+        "//libs-scala/security-evidence/tag:security-evidence-tag",
     ],
 )
