+++ ledger/sandbox-on-x/BUILD.bazel
@@ -145,7 +145,6 @@ da_scala_library(
         "//ledger/ledger-api-domain",
         "//ledger/ledger-configuration",
         "//ledger/ledger-resources",
-        "//ledger/ledger-security-evidence/lib",
         "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-integration-api:participant-integration-api-tests-lib",
@@ -157,6 +156,7 @@ da_scala_library(
         "//libs-scala/resources",
         "//libs-scala/resources-akka",
         "//libs-scala/resources-grpc",
+        "//libs-scala/security-evidence/scalatest:security-evidence-scalatest",
         "//libs-scala/security-evidence/tag:security-evidence-tag",
         "//libs-scala/timer-utils",
         "@maven//:ch_qos_logback_logback_classic",
@@ -310,7 +310,6 @@ da_scala_test_suite(
         "//ledger/ledger-resources",
         "//ledger/ledger-resources:ledger-resources-test-lib",
         "//ledger/ledger-runner-common",
-        "//ledger/ledger-security-evidence/lib",
         "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-integration-api:participant-integration-api-tests-lib",
@@ -322,6 +321,8 @@ da_scala_test_suite(
         "//libs-scala/ports",
         "//libs-scala/postgresql-testing",
         "//libs-scala/resources",
+        "//libs-scala/security-evidence/scalatest:security-evidence-scalatest",
+        "//libs-scala/security-evidence/tag:security-evidence-tag",
         "//libs-scala/timer-utils",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:io_grpc_grpc_services",
