+++ libs-scala/security-evidence/scalatest/BUILD.bazel
@@ -4,7 +4,7 @@
 load("//bazel_tools:scala.bzl", "da_scala_binary", "da_scala_library")
 
 da_scala_library(
-    name = "lib",
+    name = "security-evidence-scalatest",
     srcs = glob(["src/scala/**/*.scala"]),
     scala_deps = [
         "@maven//:org_scalactic_scalactic",
@@ -20,7 +20,7 @@ da_scala_library(
         "@maven//:io_circe_circe_generic_extras",
         "@maven//:com_lihaoyi_sourcecode",
     ],
-    tags = ["maven_coordinates=com.daml:ledger-security-evidence:__VERSION__"],
+    tags = ["maven_coordinates=com.daml:security-evidence-scalatest:__VERSION__"],
     visibility = [
         "//visibility:public",
     ],
